Add the func to the track service can calculate duration of a track

+++ Backend/track-service/src/main/java/com/app/payload/request/CreateTrackDto.java
@@ -11,7 +11,7 @@ public class CreateTrackDto {
 
 	private String title;
 	
-	private Long playlistId;
+	private Integer playlistId;
 	
 	private Integer authorId;
